Ajustes.
Erro - NullException em ClienteService

+++ src/main/java/br/senac/seguradora/projetoseguradora/controller/VeiculoController.java
@@ -56,7 +56,7 @@ public class VeiculoController {
 		}
 		
 		@DeleteMapping("/{id}")
-		public boolean excluir(@PathVariable Integer id) {
+		public boolean excluir(@PathVariable Long id) {
 			return veiculoService.excluir(id);
 		}
